refactored CartoMap; removed HelloMap from AdvancedMap
