+++ CartoMapAndroid/app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:name=".CartoMapApplication">
+        android:name=".map.CartoMapApplication">
 
         <activity android:name=".LauncherListActivity">
             <intent-filter>
@@ -25,12 +25,12 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".CartoSQLActivity" />
-        <activity android:name=".CartoUTFGridActivity" />
-        <activity android:name=".CartoTorqueActivity" />
-        <activity android:name=".CartoVectorTileActivity" />
-        <activity android:name=".CartoRasterTileActivity" />
-        <activity android:name=".CartoVisJSONActivity" />
+        <activity android:name=".map.CartoSQLActivity" />
+        <activity android:name=".map.CartoUTFGridActivity" />
+        <activity android:name=".map.CartoTorqueActivity" />
+        <activity android:name=".map.CartoVectorTileActivity" />
+        <activity android:name=".map.CartoRasterTileActivity" />
+        <activity android:name=".map.CartoVisJSONActivity" />
 
     </application>
