+++ CartoMapAndroid/app/src/main/java/com/carto/cartomap/android/map/CartoTorqueActivity.java
@@ -1,4 +1,4 @@
-package com.carto.cartomap.android;
+package com.carto.cartomap.android.map;
 
 import android.os.Bundle;
 import android.util.Log;
